requirements.txt
# Spradar Project - Python Dependencies
# Data processing and analysis
pandas==2.3.3
numpy==2.3.4

# Database connectivity
psycopg2-binary==2.9.11

# Environment management
python-dotenv==1.2.1

# Jupyter notebooks
jupyter==1.1.1
ipykernel==7.1.0

# Additional useful packages for data analysis
matplotlib-inline==0.2.1